requirements.txt
nodriver>=0.38
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
